+++ TD_React/src/AfficherDate.jsx
@@ -1,0 +1,11 @@
+
+function MainContent({jour, mois, annee, heure, minute, seconde}) {
+  return (
+    <main>
+      <p>
+      Bonjour, on est le {jour}, {mois}, {annee} et il est {heure}:{minute}:{seconde}
+      </p>
+    </main>
+  )
+}
+export default MainContent;
