+++ TD_React/src/Footer.jsx
@@ -1,0 +1,10 @@
+function Footer({annee, nom, prenom}) {
+  return (
+    <footer>
+      
+      © {annee} - {prenom}.{nom}, Tous droits réservés.
+    
+    </footer>
+  )
+}
+export default Footer;
