+++ TD_React/src/Header.jsx
@@ -1,0 +1,10 @@
+function Header({ title1, title2, logo }) {
+  return (
+    <header>
+      <img src={logo} alt="Logo" />
+      <h1>{title1}</h1>
+      <h2>{title2}</h2>
+    </header>
+  );
+}
+export default Header;
